+++ bin/commands/.errors
@@ -66,3 +66,8 @@ ZWEL0331W||The following requested component(s) are not defined in zowe.yaml: %s
 ZWEL0332E|332|Zowe YAML property zowe.runtimeDirectory is not defined. Zowe cannot validate component existence without knowing the runtime location.
 ZWEL0332W||Zowe YAML property zowe.runtimeDirectory is not defined. Zowe cannot validate component existence without knowing the runtime location.
 ZWEL0333E|333|Component '%s' is missing a valid manifest file (manifest.yaml, manifest.yml, or manifest.json) in %s.
+ZWEL0363W||%s is not %s. Skipping port ATTLS validation. 
+ZWEL0364E|364|%s is not %s. Skipping port ATTLS validation.
+ZWEL0365E|365|%s: No AT-TLS rule identified on %s:%d for user %s and jobname %s
+ZWEL0366E|366|%d Zowe port ATTLS validation(s) failed, review output for action items before running Zowe.
+
