+++ bin/commands/internal/start/prepare/index.ts
@@ -31,7 +31,7 @@ import * as zoslib from '../../../../libs/zos';
 import * as validateBind from '../../../validate/port/bind/index';
 import * as attls from '../../../../libs/attls';
 import * as validateComponentManifests from '../../../validate/components/index';
-import * as validateAttls from '../../../validate/port/attls/index';
+import * as validateAttls from '../../../libs/validateAttls';
 import * as validateCertificate from '../../../validate/certificate/index';
 
 //# This command prepares everything needed to start Zowe.
