--- bin/commands/validate/port/attls/.help
@@ -1,1 +1,0 @@
-Checks to see if the network ports requested for use by enabled Zowe components to see if ATTLS rules apply.
