--- bin/commands/validate/port/attls/.parameters
@@ -1,1 +1,0 @@
-component-name,component|o|string|||||Component name to be checked. If not provided, all enabled components are checked.
